Split code into core packages and services
