+++ .env
@@ -1,7 +1,7 @@
 PROTO_BUILD_PATH=/proto/v1
 PROTO_OUTPUT_PATH=/pkg/api/v1
 PROTO_FILENAME=upload.proto
-PORT=7501
-APP_SRC=/app
+PROTO_SRC=./core/proto-gen
+CORE_SRC=/app
+CODE_SRC=./core
 MODS=/mods
-BUILD_FILE=upload-server
