--- app/go.mod
@@ -1,8 +1,0 @@
-module github.com/omaressameldin/lazy-panda-upload-service
-
-go 1.12
-
-require (
-	github.com/golang/protobuf v1.2.0
-	google.golang.org/grpc v1.20.1
-)
