+++ core/go.mod
@@ -1,0 +1,13 @@
+module github.com/omaressameldin/lazy-panda-upload-service/core
+
+go 1.12
+
+require (
+	firebase.google.com/go v3.7.0+incompatible
+	github.com/golang/protobuf v1.2.0
+	github.com/googleapis/gax-go v2.0.2+incompatible // indirect
+	github.com/omaressameldin/lazy-panda-utils/app v0.0.0-20190526184506-6249a6f1acc2
+	golang.org/x/net v0.0.0-20190522155817-f3200d17e092
+	google.golang.org/api v0.5.0
+	google.golang.org/grpc v1.20.1
+)
