+++ docker-compose.yml
@@ -9,21 +9,19 @@ services:
     volumes:
       - type: bind
         target: ${PROTO_OUTPUT_PATH}
-        source: ./app${PROTO_OUTPUT_PATH}
+        source: ${CORE_SRC}${PROTO_OUTPUT_PATH}
       - type: bind
         target: ${PROTO_BUILD_PATH}
-        source: ./proto-gen${PROTO_BUILD_PATH}
-  lazy-panda-upload-server:
+        source: ${PROTO_SRC}${PROTO_BUILD_PATH}
+  lazy-panda-upload-core:
     build:
       context: .
       args:
-        - APP_SRC=${APP_SRC}
+        - APP_SRC=${CORE_SRC}
+        - CODE_SRC=${CODE_SRC}
         - MODS=${MODS}
-        - BUILD_FILE=${BUILD_FILE}
-    command: sh -c "cp -r ${MODS}/. ${APP_SRC} && /${BUILD_FILE} --port=${PORT}"
+    command: sh -c "cp -r ${MODS}/. ${CORE_SRC}"
     volumes:
       - type: bind
-        target: ${APP_SRC}
-        source: ./app
-    ports:
-      - "${PORT}:${PORT}"
+        target: ${CORE_SRC}
+        source: ./core
